Moving timetable scripts and funcs into own utils
